pyproject.toml
[project]
name = "ai-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.29.0",
    "litellm>=1.83.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
]
